+++ tests/conftest.py
@@ -1,6 +1,5 @@
 import json
 import logging
-from base64 import b64encode
 from collections import deque
 from functools import partial
 from itertools import count
@@ -66,15 +65,6 @@ def settings(
             }
         ),
     )
-    monkeypatch.setenv(
-        "MEX_BACKEND_API_USER_DATABASE",
-        json.dumps(
-            {
-                "read": {"Reader": "read_password"},
-                "write": {"Writer": "write_password"},
-            }
-        ),
-    )
     if is_integration_test:
         monkeypatch.setenv(
             "MEX_IDENTITY_PROVIDER",
@@ -131,20 +121,9 @@ def client_with_api_key_read_permission(client: TestClient) -> TestClient:
 
 
 @pytest.fixture
-def client_with_basic_auth_read_permission(client: TestClient) -> TestClient:
-    """Return a fastAPI test client with read permission granted by basic auth."""
-    client.headers.update(
-        {"Authorization": f"Basic {b64encode(b'Reader:read_password').decode()}"}
-    )
-    return client
-
-
-@pytest.fixture
-def client_with_basic_auth_write_permission(client: TestClient) -> TestClient:
+def client_that_is_ldap_authenticated(client: TestClient) -> TestClient:
     """Return a fastAPI test client with write permission granted by basic auth."""
-    client.headers.update(
-        {"Authorization": f"Basic {b64encode(b'Writer:write_password').decode()}"}
-    )
+    client.auth = ("Writer", "write_password")
     return client