+++ tests/testing/test_ldap.py
@@ -11,9 +11,9 @@ if TYPE_CHECKING:  # pragma: no cover
 
 @pytest.mark.integration
 def test_get_preview_person_from_login_success(
-    client_with_basic_auth_write_permission: TestClient,
+    testing_app_client_that_is_ldap_authenticated: TestClient,
 ) -> None:
-    response = client_with_basic_auth_write_permission.post(
+    response = testing_app_client_that_is_ldap_authenticated.post(
         "/v0/preview-person-from-login"
     )
     assert response.status_code == status.HTTP_200_OK, response.text
@@ -25,9 +25,9 @@ def test_get_preview_person_from_login_success(
 
 @pytest.mark.integration
 def test_search_persons_or_contact_points_in_ldap_success(
-    client_with_basic_auth_write_permission: TestClient,
+    testing_app_client_with_api_key_read_permission: TestClient,
 ) -> None:
-    response = client_with_basic_auth_write_permission.get(
+    response = testing_app_client_with_api_key_read_permission.get(
         "/v0/ldap", params={"q": "mex", "limit": 1}
     )
     assert response.status_code == status.HTTP_200_OK, response.text