+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Removed
 
 - removed Preview-item post endpoint
+- BREAKING: removed basic-auth, only API keys and ldap auth remain
 
 ### Fixed